creacion de un archivo que contiene las opciones de una tarea
