+++ src/components/TaskOptions.tsx
@@ -1,62 +1,12 @@
 import { useDispatch } from 'react-redux'
-import { MouseEventHandler } from 'react'
 import './TaskOptions.css'
-import { CaretLeftIcon, CaretRightIcon, CircleFillIcon, ClipboardIcon, TrashIcon } from './atomic/Icon'
-import toast from 'react-hot-toast'
+import { CaretLeftIcon, CaretRightIcon } from './atomic/Icon'
 import { Btn } from './atomic/Btn'
 import { COLORS_CLASS_NAME } from './atomic/colors'
 import { moveToType } from '../domainFunctions/moveTask'
-import { deleteTask, highlightTask, moveTask } from '../redux/columnsSlice'
+import { moveTask } from '../redux/columnsSlice'
 import { taskModel } from '../models/task'
-
-interface option {
-    name: string,
-    function: MouseEventHandler<HTMLButtonElement>,
-    color: COLORS_CLASS_NAME
-    icon?: Function
-}
-
-function getTaskOptions(task: taskModel): option[] {
-    const dispatch = useDispatch();
-
-    const deleteTheTask = () => {
-        dispatch(deleteTask(task))
-        toast.success('Tarea eliminada')
-    }
-
-    const highlightTheTask = () => {
-        dispatch(highlightTask(task))
-        if(!task.highlight) {
-        toast.success('Tarea resaltada')
-        }
-        else {
-        toast.success('Tarea sin resaltado')
-        }
-    }
-
-    const options = [
-        {
-          name: 'Copiar', 
-          function: () => navigator.clipboard.writeText(task.descriptionText).then(() => toast.success('Tarea copiada al portapapeles')),
-          color: COLORS_CLASS_NAME.PRIMARY,
-          icon: () => <ClipboardIcon />
-        },
-        {
-            name: 'Resaltar', 
-            function: highlightTheTask,
-            color: COLORS_CLASS_NAME.PRIMARY,
-            icon: () => <CircleFillIcon />
-          },
-        {
-          name: 'Eliminar', 
-          function: deleteTheTask,
-          color: COLORS_CLASS_NAME.DANGER,
-          icon: () => <TrashIcon />
-        },
-    ]
-
-    return options
-}
+import { getTaskOptions } from '../taskOptions'
 
 interface TaskOptionsProps {
     task: taskModel
